2017_Transformer/Transformer/Optim.py
"""学习率调度优化器的包装类"""

import numpy as np


class ScheduledOptim:
    """学习率调度的简单包装类"""

    def __init__(self, optimizer, lr_mul, d_model, n_warmup_steps):
        self._optimizer = optimizer
        self.lr_mul = lr_mul
        self.d_model = d_model
        self.n_warmup_steps = n_warmup_steps
        self.n_steps = 0

    def step_and_update_lr(self):
        "使用内部优化器进行步骤更新"
        self._update_learning_rate()
        self._optimizer.step()

    def zero_grad(self):
        "使用内部优化器清零梯度"
        self._optimizer.zero_grad()

    def _get_lr_scale(self):
        d_model = self.d_model
        n_steps, n_warmup_steps = self.n_steps, self.n_warmup_steps
        return (d_model**-0.5) * min(
            n_steps ** (-0.5), n_steps * n_warmup_steps ** (-1.5)
        )

    def _update_learning_rate(self):
        """每步学习率调度"""

        self.n_steps += 1
        lr = self.lr_mul * self._get_lr_scale()

        for param_group in self._optimizer.param_groups:
            param_group["lr"] = lr
