2017_Transformer/Transformer/requirements.txt
dill==0.3.3
msgpack-numpy==0.4.7.1
msgpack==1.0.2
torch==1.3.1
torchtext==0.4.0
spacy==2.3.5
tensorboard==1.14.0
tensorflow==1.14.0
terminado==0.9.2
tqdm==4.56.0
